Improve build-lib package
