+++ utils/build-lib/package.json
@@ -24,7 +24,6 @@
   "dependencies": {
     "mri": "^1.1.4",
     "rollup": "^1.14.5",
-    "rollup-plugin-buble": "^0.19.6",
     "rollup-plugin-commonjs": "^10.0.0",
     "rollup-plugin-node-resolve": "^5.0.1",
     "rollup-plugin-typescript": "^1.0.1",
